And update on the log activity.  This avoids crashes, but has a more fundamental problem.
The PlayGameAct activity has lots of state.  Many classes worth of state.  Switching to the DisplayLogAct activity loses that state because it is not recovered when shifting back to the PlayGameAct.

I could have the 'Show Log' button click save state and the onResume method restore that state.  In some ways that would be similar to what I'll need to do about saving for backgrounding the app, so that might be the way to go.  A more direct way for right now would be change displaying the log from being a separate activity to being a display panel on top of the main screen which is shown and hidden.  I'll probably do that next as being simpler.