Added the ability to run in "automatic" mode.
This is just an alternative starting action which has defaults for a 4 machine player game.  A run configuration can be used to run/debug from either than standard Intro1Act or the new AutoBuyoutAct.

The run doesn't show anything on the emulator until the final end of game screen.  Having it show an initial screen and wait for a human button push would be counter productive since the idea is ultimately to be able to make many automatic runs for testing machine players.